git-svn-id: https://fileserver.centit.com/svn/centit/products/centit-meta-form@145366 11746a9b-a569-1649-8427-d29ac4cc1c6a

+++ src/main/java/com/centit/metaform/controller/MetaFormController.java
@@ -1,18 +1,19 @@
 package com.centit.metaform.controller;
 
 import java.sql.SQLException;
-import java.util.HashMap;
 import java.util.Map;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.context.ContextLoaderListener;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
@@ -21,11 +22,8 @@ import com.centit.framework.core.common.ResponseData;
 import com.centit.framework.core.controller.BaseController;
 import com.centit.framework.core.dao.PageDesc;
 import com.centit.metaform.formaccess.ModelFormService;
+import com.centit.metaform.formaccess.OperationEvent;
 import com.centit.metaform.formaccess.impl.JdbcModelRuntimeContext;
-import com.centit.support.algorithm.DatetimeOpt;
-import com.centit.support.algorithm.NumberBaseOpt;
-import com.centit.support.algorithm.StringBaseOpt;
-import com.centit.support.database.metadata.TableField;
 
 @Controller
 @RequestMapping("/metaform/formaccess")
@@ -34,6 +32,8 @@ public class MetaFormController  extends BaseController{
 
 	@Resource(name="jdbcModelFormService")
 	private ModelFormService formService;
+	
+	
 	
 	@RequestMapping(value = "/{modelCode}/list",method = RequestMethod.GET)
 	public void list(@PathVariable String modelCode,boolean addMeta, PageDesc pageDesc ,HttpServletRequest request, HttpServletResponse response) {
@@ -52,33 +52,18 @@ public class MetaFormController  extends BaseController{
 	
 	
 	
-	@RequestMapping(value = "/{modelCode}/get",method = RequestMethod.GET)
+	@RequestMapping(value = "/{modelCode}/view",method = RequestMethod.GET)
 	public void view(@PathVariable String modelCode, boolean addMeta, HttpServletRequest request, HttpServletResponse response) {
     	JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
-    	Map<String,Object> jo = new HashMap<>();
-    	for(String pk: rc.getTableInfo().getPkColumns()){
-    		TableField pkp = rc.getTableInfo().findFieldByColumn(pk);
-    		Object pv = request.getParameter(pkp.getPropertyName());
-    		switch(pkp.getJavaType()){
-			case "Date":
-			case "Timestamp":
-    			jo.put(pkp.getPropertyName(), DatetimeOpt.castObjectToDate(pv));
-    			break;
-			case "Long":
-    			jo.put(pkp.getPropertyName(), NumberBaseOpt.castObjectToLong(pv));
-    			break;
-			case "Double":
-    			jo.put(pkp.getPropertyName(), NumberBaseOpt.castObjectToDouble(pv));
-    			break;
-			default:
-    			jo.put(pkp.getPropertyName(), StringBaseOpt.objectToString(pv));
-    			break;
-    		}
-    	}
+    	Map<String,Object> jo = rc.fetchPkFromRequest(request);    	
 		ResponseData resData = new ResponseData();
-		resData.addResponseData("fields", rc.getFormDefine("view").getFilters());
 		resData.addResponseData("obj", formService.getObjectByProperties(rc, jo));
-		rc.close();
+		
+		rc.close();		
+		if(addMeta){
+        	resData.addResponseData("formModel",  rc.getFormDefine("view")); 
+        }
+		
 		JsonResultUtils.writeResponseDataAsJson(resData, response);
 	}
 	
@@ -91,14 +76,17 @@ public class MetaFormController  extends BaseController{
 	}
 	
 	@RequestMapping(value = "/{modelCode}/create",method = RequestMethod.GET)
-	public void create(@PathVariable String modelCode, @RequestBody String jsonStr,
-			boolean createPk, HttpServletRequest request, HttpServletResponse response) {
-		JSONObject jo = JSON.parseObject(jsonStr);
+	public void create(@PathVariable String modelCode,
+			boolean addMeta, HttpServletRequest request, HttpServletResponse response) {
     	JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
     	try {
-			formService.saveNewObject(rc, jo);
-			rc.commitAndClose();
-			JsonResultUtils.writeSuccessJson(response);
+    		ResponseData resData = new ResponseData();
+    		resData.addResponseData("obj", formService.createInitialObject(rc));    		
+    		rc.close();
+			if(addMeta){
+	        	resData.addResponseData("formModel",  rc.getFormDefine("create")); 
+	        }			
+			JsonResultUtils.writeResponseDataAsJson(resData, response);
 		} catch (SQLException e) {
 			JsonResultUtils.writeErrorMessageJson(e.getMessage(), response);
 			rc.rollbackAndClose();
@@ -107,12 +95,13 @@ public class MetaFormController  extends BaseController{
 	}
 	
 	@RequestMapping(value = "/{modelCode}/createpk",method = RequestMethod.GET)
-	public void createPk(@PathVariable String modelCode, @RequestBody String jsonStr,  HttpServletRequest request, HttpServletResponse response) {
-		JSONObject jo = JSON.parseObject(jsonStr);
+	public void createPk(@PathVariable String modelCode,
+			HttpServletRequest request, HttpServletResponse response) {
     	JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
     	try {
-			formService.saveNewObject(rc, jo);
-			rc.commitAndClose();
+    		ResponseData resData = new ResponseData();
+    		resData.addResponseData("pk", formService.createNewPk(rc));    	
+			rc.close();
 			JsonResultUtils.writeSuccessJson(response);
 		} catch (SQLException e) {
 			JsonResultUtils.writeErrorMessageJson(e.getMessage(), response);
@@ -120,16 +109,70 @@ public class MetaFormController  extends BaseController{
 			e.printStackTrace();
 		}		
 	}
+	/**
+	 * 
+	 * @param rc
+	 * @param jo
+	 * @param eventType
+	 * @param response
+	 * @return 
+	 * @throws Exception
+	 */
+	private int runOperationEvent(JdbcModelRuntimeContext rc, JSONObject jo , 
+			String eventType , HttpServletResponse response ) throws Exception{
+		String eventBeanName = rc.getMetaFormModel().getExtendOptBean();
+		if(StringUtils.isBlank(eventBeanName))
+				return 0;		
+		OperationEvent optEvent = 
+	                ContextLoaderListener.getCurrentWebApplicationContext().
+	                getBean(eventBeanName,  OperationEvent.class);
+		if(optEvent==null)
+			return -1;
+		switch(eventType){
+		case "beforeSave":
+			return optEvent.beforeSave(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		case "afterSave":
+			return optEvent.afterSave(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		case "beforeUpdate":
+			return optEvent.beforeUpdate(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		case "afterUpdate":
+			return optEvent.afterUpdate(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		case "beforeDelete":
+			return optEvent.beforeDelete(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		case "afterDelete":
+			return optEvent.afterDelete(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		case "beforeSubmit":
+			return optEvent.beforeSubmit(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		case "afterSubmit":
+			return optEvent.afterSubmit(rc, jo, 
+					rc.getMetaFormModel().getExtendOptBeanParam(), response);
+		default:
+			return 0;
+		}
+	}
 	
 	@RequestMapping(value = "/{modelCode}/save",method = RequestMethod.POST)
-	public void saveNew(@PathVariable String modelCode, @RequestBody String jsonStr,  HttpServletRequest request, HttpServletResponse response) {
+	public void saveNew(@PathVariable String modelCode, @RequestBody String jsonStr, 
+			HttpServletRequest request, HttpServletResponse response) {
 		JSONObject jo = JSON.parseObject(jsonStr);
     	JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
     	try {
-			formService.saveNewObject(rc, jo);
-			rc.commitAndClose();
-			JsonResultUtils.writeSuccessJson(response);
-		} catch (SQLException e) {
+    		int n = runOperationEvent(rc, jo, "beforeSave", response);
+    		if( n<=0 ){
+    			formService.saveNewObject(rc, jo);
+     			n = runOperationEvent(rc, jo, "afterSave", response);
+    		}
+    		rc.commitAndClose();			
+			if(n<=1)
+				JsonResultUtils.writeSuccessJson(response);
+		} catch (Exception e) {
 			JsonResultUtils.writeErrorMessageJson(e.getMessage(), response);
 			rc.rollbackAndClose();
 			e.printStackTrace();
@@ -137,19 +180,19 @@ public class MetaFormController  extends BaseController{
 	}
 	
 	@RequestMapping(value = "/{modelCode}/edit",method = RequestMethod.GET)
-	public void edit(@PathVariable String modelCode, @RequestBody String jsonStr,  HttpServletRequest request, HttpServletResponse response) {
+	public void edit(@PathVariable String modelCode,boolean addMeta,
+			HttpServletRequest request, HttpServletResponse response) {
         
-		JSONObject jo = JSON.parseObject(jsonStr);
-    	JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
-    	try {
-			formService.updateObject(rc, jo);
-			rc.commitAndClose();
-			JsonResultUtils.writeSuccessJson(response);
-		} catch (SQLException e) {
-			JsonResultUtils.writeErrorMessageJson(e.getMessage(), response);
-			rc.rollbackAndClose();
-			e.printStackTrace();
-		}
+		JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
+    	Map<String,Object> jo = rc.fetchPkFromRequest(request);    	
+		ResponseData resData = new ResponseData();
+		resData.addResponseData("obj", formService.getObjectByProperties(rc, jo));		
+		rc.close();		
+		if(addMeta){
+        	resData.addResponseData("formModel",  rc.getFormDefine("edit")); 
+        }
+		
+		JsonResultUtils.writeResponseDataAsJson(resData, response);
 	}
 	
 	
@@ -159,10 +202,15 @@ public class MetaFormController  extends BaseController{
 		JSONObject jo = JSON.parseObject(jsonStr);
     	JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
     	try {
-			formService.updateObject(rc, jo);
-			rc.commitAndClose();
-			JsonResultUtils.writeSuccessJson(response);
-		} catch (SQLException e) {
+			int n = runOperationEvent(rc, jo, "beforeUpdate", response);
+    		if( n<=0 ){
+    			formService.updateObject(rc, jo);
+     			n = runOperationEvent(rc, jo, "afterUpdate", response);
+    		}
+    		rc.commitAndClose();			
+			if(n<=1)
+				JsonResultUtils.writeSuccessJson(response);
+		} catch (Exception e) {
 			JsonResultUtils.writeErrorMessageJson(e.getMessage(), response);
 			rc.rollbackAndClose();
 			e.printStackTrace();
@@ -174,10 +222,15 @@ public class MetaFormController  extends BaseController{
 		JSONObject jo = JSON.parseObject(jsonStr);
     	JdbcModelRuntimeContext rc = (JdbcModelRuntimeContext)formService.createRuntimeContext(modelCode);
     	try {
-			formService.deleteObjectById(rc, jo);
-			rc.commitAndClose();
-			JsonResultUtils.writeSuccessJson(response);
-		} catch (SQLException e) {
+			int n = runOperationEvent(rc, jo, "beforeDelete", response);
+    		if( n<=0 ){
+    			formService.deleteObjectById(rc, jo);
+     			n = runOperationEvent(rc, jo, "afterDelete", response);
+    		}
+    		rc.commitAndClose();			
+			if(n<=1)
+				JsonResultUtils.writeSuccessJson(response);
+		} catch (Exception e) {
 			JsonResultUtils.writeErrorMessageJson(e.getMessage(), response);
 			rc.rollbackAndClose();
 			e.printStackTrace();
